Update test_wordfrequency.py
I might have missed the point of this test, but it should look for an empty string - not an empty list! At least, I believe that is what .strip() should be returning after being asked to remove all characters from a string. 

Also, small.txt and voluspaa.txt are not found in this folder.